Add confirmation for deleting all notes

+++ src/commands/delete.rs
@@ -1,3 +1,4 @@
+use std::io::{self, Write};
 
 use rusqlite::{Connection, Result};
 
@@ -7,14 +8,29 @@ use crate::db::delete_note::delete_note;
 
 pub fn delete(cmd: DeleteCommand, conn: &Connection) -> Result<()> {
     if cmd.all {
-        let rows = delete_all_notes(conn)?;
-        if rows > 0 {
-            println!("All notes deleted!");
-        } else {
-            println!("No notes to delete")
-        }
+        let mut confirm = String::new();
+        print!("Delete all notes? y/n: ");
+        io::stdout().flush().expect("Failed to flush stdout");
+        
+        io::stdin()
+            .read_line(&mut confirm)
+            .expect("Failed to read input");
 
-        return Ok(())
+        confirm = confirm.trim().to_lowercase();
+    
+        if confirm == "y" || confirm == "yes" {
+            let rows = delete_all_notes(conn)?;
+            if rows > 0 {
+                println!("All notes deleted!");
+            } else {
+                println!("No notes to delete")
+            }
+
+            return Ok(())
+        } else {
+            println!("Note deletion stopped");
+            std::process::exit(0);
+        }
     } 
 
     if cmd.title.is_some() && cmd.id.is_some() {
